ipn/ipnlocal: extract ping implementation into its own file
There’s enough code to pull this out into its own ping.go file, since
local.go contains a lot of stuff.

The next commit will add a Pinger interface to this file, so I am
breaking out this refactor to make identifying the changes and
reviewing the actual patch easier.
